Set up TS config and transform Icon to TS
